Create reusable `DismissButton`
