+++ apps/store/src/components/QuickAdd/DismissButton.tsx
@@ -1,0 +1,20 @@
+import { datadogRum } from '@datadog/browser-rum'
+import { useTranslation } from 'next-i18next'
+import { Button } from 'ui'
+import { useShowQuickAdd } from './useShowQuickAdd'
+
+export const DismissButton = () => {
+  const { t } = useTranslation('cart')
+  const [, setShow] = useShowQuickAdd()
+
+  const handleDismiss = () => {
+    datadogRum.addAction('Quick Add Hide')
+    setShow(false)
+  }
+
+  return (
+    <Button size="medium" variant="ghost" onClick={handleDismiss}>
+      {t('QUICK_ADD_DISMISS')}
+    </Button>
+  )
+}
